Fix visits not being logged
The UI for user visits is showing as zero. This is because we don't
actually log the visit! This commit adds a middleware that simply logs a
visit with Visitor::log() on any web request.

We are using a hardcoded base value when showing the count so that the
user count is relatively accurate. This was taken from cloudfare at the
time of the fix.