Fix external link clicks being lost
The clicks on links to the external websites were not being counted.
This was caused by two issues:

1st is that we were passing the org id to the route, but the
Organization model defines the slug as the routeKeyName. This meant it
would always 404 as it was using the db id to search by slug!

2nd issue was to do with how the website property on thge Organization
was being access in the OrganizationClicksController. It was referenced
as a property instead of a function, which was throwing
`App\Organization::website must return a relationship
instance.` Changing it to ->website() to match how it's used in
OrganizationController@show fixes this.